# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project_Alan API documentation example",
        "contact": {
            "email": "aldaberdiev99@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Display all Categories",
                "operationId": "categoriessAll",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryShowRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Store category data",
                "description": "Store category data",
                "operationId": "categoryStore",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of Category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "The slug of Category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "photo",
                        "in": "query",
                        "description": "The photo of category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "icon",
                        "in": "query",
                        "description": "The icon of category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass action credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "price",
                                    "title"
                                ],
                                "properties": {
                                    "price": {
                                        "type": "integer",
                                        "example": "20"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "About actions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Fill all required fields"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by ID",
                "operationId": "categoriesGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryShowRequest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category by ID",
                "operationId": "categoriesUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActionStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryShowRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category by ID",
                "operationId": "categoriesDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of action",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/categories/{id}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Display all category products",
                "operationId": "categoriesProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of action",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductShowRequest"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ActionStoreRequest": {
                "title": "Action post example",
                "description": "Some simple request createad as action",
                "properties": {
                    "price": {
                        "title": "price",
                        "description": "Price of action for storring",
                        "type": "integer",
                        "example": "55"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title for storring",
                        "type": "string",
                        "example": "About action"
                    }
                },
                "type": "object"
            },
...
            "CategoryShowRequest": {
                "title": "Category get example",
                "description": "Some simple example",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of the category",
                        "type": "integer",
                        "example": "Fruits"
                    },
                    "slug": {
                        "title": "slug",
                        "description": "Slug for category",
                        "type": "string",
                        "example": "About action"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "Photo of category",
                        "type": "string",
                        "example": "https://www.google.com/search?q=fruits&sxsrf=ALeKk00am-a2RSceVhR9xhnmtlTf6hyi4A:1619097112058&tbm=isch&source=iu&ictx=1&fir=Iqrn9zwUMx_LVM%252CV1GXApsZb-s1oM%252C_&vet=1&usg=AI4_-kQn1ZS8qcz8wpqJ5SAbYI3mivconA&sa=X&ved=2ahUKEwjs-OHh9pHwAhXSpIsKHfgoB5kQ9QF6BAgSEAE#imgrc=Iqrn9zwUMx_LVM"
                    },
                    "category_icon": {
                        "title": "Category_icon",
                        "description": "Category icon",
                        "type": "string",
                        "example": "https://www.google.com/search?q=fruits&sxsrf=ALeKk00am-a2RSceVhR9xhnmtlTf6hyi4A:1619097112058&tbm=isch&source=iu&ictx=1&fir=Iqrn9zwUMx_LVM%252CV1GXApsZb-s1oM%252C_&vet=1&usg=AI4_-kQn1ZS8qcz8wpqJ5SAbYI3mivconA&sa=X&ved=2ahUKEwjs-OHh9pHwAhXSpIsKHfgoB5kQ9QF6BAgSEAE#imgrc=Iqrn9zwUMx_LVM"
                    },
                    "created_at": {
                        "title": "Created_at",
                        "description": "Created time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated_at",
                        "description": "Updated time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    }
                },
                "type": "object"
            },
...
            "ProductShowRequest": {
                "title": "Product get example",
                "description": "Some simple example",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Unique ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of the product",
                        "type": "integer",
                        "example": "Banana"
                    },
                    "price": {
                        "title": "price",
                        "description": "Price of products",
                        "type": "integer",
                        "example": "60"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "Photo of category",
                        "type": "string",
                        "example": "https://www.google.com/search?q=fruits&sxsrf=ALeKk00am-a2RSceVhR9xhnmtlTf6hyi4A:1619097112058&tbm=isch&source=iu&ictx=1&fir=Iqrn9zwUMx_LVM%252CV1GXApsZb-s1oM%252C_&vet=1&usg=AI4_-kQn1ZS8qcz8wpqJ5SAbYI3mivconA&sa=X&ved=2ahUKEwjs-OHh9pHwAhXSpIsKHfgoB5kQ9QF6BAgSEAE#imgrc=Iqrn9zwUMx_LVM"
                    },
                    "created_at": {
                        "title": "Created_at",
                        "description": "Created time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated_at",
                        "description": "Updated time",
                        "type": "string",
                        "example": "2021-04-22T12:11:09.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}